+++ src/router/index.js
@@ -7,8 +7,8 @@ import SignUp from "@/components/user/signup/SignUp.vue";
 import EmailSignUp from "@/components/user/signup/child/EmailSignUp.vue";
 import SignIn from "@/components/user/signin/SignIn.vue";
 
-import Notice from "@/components/user/notice/Notice.vue"
-import Mypage from "@/components/user/mypage/Mypage.vue"
+import Notice from "@/components/user/notice/Notice.vue";
+import Keyword from "@/components/user/mypage/Keyword.vue";
 
 Vue.use(VueRouter);
 
@@ -49,6 +49,11 @@ const routes = [
         path: "mypage",
         name: "Mypage",
         component: Mypage,
+      },
+      {
+        path: "keyword",
+        name: "Keyword",
+        component: Keyword,
       },
     ],
   },
